+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 Notable changes to this project will be documented in this file
 
-- [X] Add in User Authentication using JWT and password hashing using Bcrypt - 6/30/22
+- [X] Update bcrypt import causing password error - 7/6/22
+- [X] Add in User Authentication using JWT and password hashing using bcrypt - 6/30/22
 - [X] Basic CRUD functionality working for User - 6/29/22
 - [X] Add User entity - 6/29/22
 - [X] Connect to Postgres DB - 6/24/22